# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Marketplace Nigeria API Documentation'
  description: 'Complete REST API documentation for Marketplace Nigeria platform. This API allows you to manage adverts, users, messages, payments, and more.'
  version: '3.0'
servers:
  -
    url: 'http://127.0.0.1:8030'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can obtain your authentication token by logging in through the `/api/login` endpoint. The token should be included in the Authorization header as a Bearer token for all authenticated requests.'
security:
  -
    default: []
paths:
...
  '/api/adverts/{advertId}/boost-info':
    get:
      summary: ''
      operationId: getApiAdvertsAdvertIdBoostInfo
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Boosts
      security: []
    parameters:
      -
        in: path
        name: advertId
        description: ''
        example: deserunt
        required: true
        schema:
          type: string
  '/api/adverts/{advertId}/boost':
    post:
      summary: ''
      operationId: postApiAdvertsAdvertIdBoost
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Boosts
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                duration_days:
                  type: integer
                  description: 'Must be at least 1. Must not be greater than 30.'
                  example: 24
                  nullable: true
      security: []
    parameters:
      -
        in: path
        name: advertId
        description: ''
        example: commodi
        required: true
        schema:
          type: string
  '/api/adverts/{advertId}/boost-status':
    get:
      summary: ''
      operationId: getApiAdvertsAdvertIdBoostStatus
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Boosts
      security: []
    parameters:
      -
        in: path
        name: advertId
        description: ''
        example: vel
        required: true
        schema:
          type: string
  '/api/boosts/{boostId}':
    get:
      summary: ''
      operationId: getApiBoostsBoostId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Boosts
      security: []
    delete:
      summary: ''
      operationId: deleteApiBoostsBoostId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Boosts
      security: []
    parameters:
      -
        in: path
        name: boostId
        description: ''
        example: inventore
        required: true
        schema:
          type: string
  '/api/boosts/{boostId}/upload-proof':
    post:
      summary: ''
      operationId: postApiBoostsBoostIdUploadProof
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Boosts
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                payment_proof:
                  type: string
                  format: binary
                  description: 'Must be a file. Must not be greater than 12048 kilobytes.'
              required:
                - payment_proof
      security: []
    parameters:
      -
        in: path
        name: boostId
        description: ''
        example: id
        required: true
        schema:
          type: string
...